+++ src/main/java/ru/skypro/courseWork/entity/Comment.java
@@ -11,7 +11,7 @@ public class Comment {
 
     private String authorFirstName;
 
-    private Integer createAt;
+    private Long createAt;
 
     private Integer pk;
